pyproject.toml
[project]
name = "gowith_be"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "STCN-2", email = "stcn2@smart-teach.cn" }]
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "python-decouple>=3.8",
    "psycopg2-binary>=2.9.10",
    "casdoor>=1.32.0",
    "drf-spectacular[sidecar]>=0.28.0",
    "openai>=1.97.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "django-filter>=25.1",
    "flower>=2.0.1",
    "django-celery-beat>=2.8.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "GPL3" }


[tool.pdm]
distribution = false
